+++ src/refactoring/Providers.tsx
@@ -2,7 +2,8 @@ import { ReactNode } from "react";
 import { ProductProvider } from "./contexts/ProductContext";
 import { CouponProvider } from "./contexts/CouponContext";
 import { CartProvider } from "./contexts/CartContext";
-import { initialCoupons, initialProducts } from "./constants";
+import { initialProducts } from "./constants/initial-data/products";
+import { initialCoupons } from "./constants/initial-data/coupons";
 
 export const Providers = ({ children }: { children: ReactNode }) => {
   return (
