--- src/refactoring/constants/index.ts
@@ -1,2 +1,0 @@
-export * from "./initialProducts";
-export * from "./initialCoupons";
