+++ src/refactoring/constants/initial-data/products.ts
@@ -1,4 +1,4 @@
-import { Product } from "../../types";
+import { Product } from "../../../types";
 
 export const initialProducts: Product[] = [
   {
