+++ src/refactoring/hooks/entities/useCart.ts
@@ -1,7 +1,7 @@
 // useCart.ts
 import { useState } from "react";
-import { CartItem, Coupon, Product } from "../../types";
-import { calculateCartTotal, updateCartItemQuantity } from "../models/cart";
+import { CartItem, Coupon, Product } from "../../../types";
+import { calculateCartTotal, updateCartItemQuantity } from "../../models/cart";
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
